+++ src/main/java/me/khajiitos/jackseconomy/JacksEconomy.java
@@ -1,27 +1,39 @@
 package me.khajiitos.jackseconomy;
 
 import com.mojang.logging.LogUtils;
+import io.netty.buffer.Unpooled;
 import me.khajiitos.jackseconomy.config.ClientConfig;
 import me.khajiitos.jackseconomy.config.Config;
 import me.khajiitos.jackseconomy.create.CreateCheck;
 import me.khajiitos.jackseconomy.create.CreateStressProvider;
 import me.khajiitos.jackseconomy.curios.CuriosCheck;
 import me.khajiitos.jackseconomy.curios.CuriosHandler;
+import me.khajiitos.jackseconomy.data.AdminShopColorManager;
+import me.khajiitos.jackseconomy.data.price.PriceManager;
 import me.khajiitos.jackseconomy.gamestages.GameStagesManager;
 import me.khajiitos.jackseconomy.init.*;
 import me.khajiitos.jackseconomy.listener.ConfigEventListeners;
 import me.khajiitos.jackseconomy.listener.OtherEventListeners;
+import me.khajiitos.jackseconomy.packet.PricesInfoPacket;
+import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.MinecraftServer;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.network.ICustomPacket;
+import net.minecraftforge.network.NetworkEvent;
+import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Mod(JacksEconomy.MOD_ID)
 public class JacksEconomy {
@@ -33,7 +45,6 @@ public class JacksEconomy {
         MinecraftForge.EVENT_BUS.register(this);
         MinecraftForge.EVENT_BUS.register(new ConfigEventListeners());
         MinecraftForge.EVENT_BUS.register(new OtherEventListeners());
-        MinecraftForge.EVENT_BUS.addListener(JacksEconomy::onRegisterCommands);
 
         AdminShopCommand.init(MinecraftForge.EVENT_BUS);
 
@@ -62,7 +73,8 @@ public class JacksEconomy {
         GameStagesManager.init();
     }
 
-    public static void onRegisterCommands(RegisterCommandsEvent e) {
+    @SubscribeEvent
+    public void onRegisterCommands(RegisterCommandsEvent e) {
         EconomyCommand.register(e.getDispatcher());
     }
 }