+++ src/main/java/me/khajiitos/jackseconomy/init/EconomyCommand.java
@@ -136,7 +136,7 @@ public class EconomyCommand {
 			}
 
 			PriceManager.save();
-			PriceManager.sendDataToPlayers();
+			PriceManager.sendDataToPlayers(false);
 
 			return 0;
 		}
@@ -173,7 +173,7 @@ public class EconomyCommand {
 			}
 
 			PriceManager.save();
-			PriceManager.sendDataToPlayers();
+			PriceManager.sendDataToPlayers(false);
 
 			return 0;
 		}
@@ -217,7 +217,7 @@ public class EconomyCommand {
 			}
 
 			PriceManager.save();
-			PriceManager.sendDataToPlayers();
+			PriceManager.sendDataToPlayers(false);
 
 			return 0;
 		}
@@ -261,7 +261,7 @@ public class EconomyCommand {
 			}
 
 			PriceManager.save();
-			PriceManager.sendDataToPlayers();
+			PriceManager.sendDataToPlayers(false);
 
 			return 0;
 		}