+++ src/main/java/me/khajiitos/jackseconomy/listener/OtherEventListeners.java
@@ -26,7 +26,7 @@ public class OtherEventListeners {
     @SubscribeEvent
     public void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent e) {
         if (e.getEntity() instanceof ServerPlayer serverPlayer) {
-            Packets.sendToClient(serverPlayer, new PricesInfoPacket(PriceManager.toTag(false), PriceManager.toTag(true)));
+            PriceManager.sendDataToPlayer(serverPlayer, true);
             AdminShopColorManager.updatePlayer(serverPlayer);
         }
     }